refactor(BinaryReader): moved out binaryReader to utils
