+++ packages/io-utils/BinaryReader.js
