+++ packages/stl-deserializer/package.json
@@ -33,6 +33,7 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@jscad/csg": "0.3.1"
+    "@jscad/csg": "0.3.1",
+    "@jscad/io-utils": "0.0.4"
   }
 }
